factorize get_year_month_or_none to itou.archive.utils
prepare reuse in anonymize_professionals command
