+++ itou/archive/utils.py
@@ -1,4 +1,7 @@
+import datetime
+
 from django.db import models
+from django.utils import timezone
 
 from itou.users.models import User
 
@@ -13,3 +16,13 @@ def get_filter_kwargs_on_user_for_related_objects_to_check():
         for obj in User._meta.related_objects
         if getattr(obj, "on_delete", None) and obj.on_delete != models.CASCADE
     }
+
+
+def get_year_month_or_none(date=None):
+    if not date:
+        return None
+
+    if isinstance(date, datetime.datetime):
+        return timezone.localdate(date).replace(day=1)
+
+    return date.replace(day=1)
