+++ tests/archive/tests_utils.py
@@ -1,6 +1,11 @@
-from django.db import models
+import datetime
 
-from itou.archive.utils import get_filter_kwargs_on_user_for_related_objects_to_check
+import pytest
+import pytz
+from django.db import models
+from django.utils import timezone
+
+from itou.archive.utils import get_filter_kwargs_on_user_for_related_objects_to_check, get_year_month_or_none
 from itou.users.models import User
 
 
@@ -19,3 +24,16 @@ class TestRelatedObjectsConsistency:
             if getattr(obj, "on_delete", None) and obj.on_delete == models.CASCADE
         ]
         assert user_related_objects == snapshot(name="user_related_objects_deleted_on_cascade")
+
+
+@pytest.mark.parametrize(
+    "date_input, expected_output",
+    [
+        (timezone.make_aware(datetime.datetime(2023, 10, 15)), datetime.date(2023, 10, 1)),
+        (timezone.make_aware(datetime.datetime(2024, 8, 31, 23, 0, 0), pytz.utc), datetime.date(2024, 9, 1)),
+        (datetime.date(2023, 10, 15), datetime.date(2023, 10, 1)),
+        (None, None),
+    ],
+)
+def test_get_year_month_or_none(date_input, expected_output):
+    assert get_year_month_or_none(date_input) == expected_output
